requirements.txt
# Computer Vision & Machine Learning
opencv-python<4.10.0
opencv-contrib-python<4.10.0  # For BEBLID descriptor
ultralytics>=8.3.0            # YOLO11
torch>=2.0.0
torchvision>=0.15.0
easyocr>=1.7.1
roboflow>=1.1.0               # For downloading pre-trained models

# Numerical Computing
numpy<2.0.0
scipy==1.11.3
filterpy>=1.4.5

# API & Web
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.20
websockets==12.0
aiohttp==3.9.1

# Price Data APIs
requests==2.31.0
python-dotenv==1.0.0
tenacity>=8.2.0              # Retry with exponential backoff
beautifulsoup4>=4.12.0       # TCGPlayer scraper fallback
lxml>=5.0.0                  # Fast HTML parsing

# Utilities
pillow==10.1.0
pyyaml==6.0.1
tqdm==4.66.1
pandas>=2.0.0  # For training notebooks and data analysis

# Performance Optimization
numba==0.58.1  # JIT compilation for feature matching

# Database
chromadb>=0.4.18